requirements.txt
pillow
fastapi
uvicorn[standard]
websockets
aiohttp[speedups]

twisted
python-xlib
